move autoencoder_kl_qwen_image to module v2
