+++ max/python/max/pipelines/architectures/autoencoders/__init__.py
@@ -13,3 +13,4 @@
 
 from .autoencoder_kl import AutoencoderKLModel
 from .autoencoder_kl_flux2 import AutoencoderKLFlux2Model
+from .autoencoder_kl_qwen_image import AutoencoderKLQwenImageModel
