+++ max/python/max/pipelines/architectures/autoencoders/autoencoder_kl_qwen_image.py
